Fixed wrong type.

+++ api/include/redpitaya/lockbox.h
@@ -1436,10 +1436,10 @@ int rp_PIDGetKi(rp_pid_t pid, float *ki);
 int rp_PIDSetKd(rp_pid_t pid, uint32_t kd);
 int rp_PIDGetKd(rp_pid_t pid, uint32_t *kd);
 
-int rp_PIDSetKii(rp_pid_t pid, uint32_t kii);
-int rp_PIDGetKii(rp_pid_t pid, uint32_t *kii);
-int rp_PIDSetKg(rp_pid_t pid, uint32_t kg);
-int rp_PIDGetKg(rp_pid_t pid, uint32_t *kg);
+int rp_PIDSetKii(rp_pid_t pid, float kii);
+int rp_PIDGetKii(rp_pid_t pid, float *kii);
+int rp_PIDSetKg(rp_pid_t pid, float kg);
+int rp_PIDGetKg(rp_pid_t pid, float *kg);
 
 /*
  * Enable or disable the integrator reset of the specified PID.
